pyproject.toml
[project]
name = "ocr-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
